example setpoint and measurements should be 16 bit

+++ examples/DCMotorControllerWithAutoTune/DCMotorControllerWithAutoTune.ino
@@ -71,8 +71,8 @@ const int ONBOARD_LED =	  13;
  *---------------------------------------------------------------------
  */
 
-volatile int32_t positionMeasurement = 0;
-volatile int32_t positionSetpoint = 0; 
+volatile int16_t positionMeasurement = 0;
+volatile int16_t positionSetpoint = 0; 
 uint32_t pidExecutionTimer = 0;
 uint32_t pidExecutionIntervalUs = 1000000 / PID_FREQ_HZ;
